Improved sales_metrics table

+++ ProductService/src/main/java/com/product/service/dto/productDetail/ProductDto.java
@@ -20,7 +20,7 @@ public class ProductDto {
     private String name;
     private String description;
     private Double price;
-    private Integer soldLastMonthCount;
+    private Long soldLastMonthCount;
     private SellerSimpleDto seller;
     private CategoryDto category;
     private Integer inventoryCount;
